+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## [Unreleased]
 
 (feature/created-repos-pagination)
-### Changed
+#### Added
+- components: UserActivity: load more repositories on demand
+- api: user: contributions/repository-contributions endpoint
+#### Changed
 - components: ActivityOverview: display percentages with two digits after decimal point
 
 (feature/repos-contributed-pagination)